requirements.txt
streamlit==1.36.0
pandas==2.2.2
openpyxl==3.1.5
pypdf==4.3.1
python-dotenv==1.0.1

openai==1.40.0
anthropic==0.34.2
google-generativeai==0.7.2
requests==2.32.3

